+++ inst/examples/shipping/R/discount.R
@@ -1,0 +1,3 @@
+apply_discount <- function(price) {
+  price - 1
+}